+++ test/uk/gov/hmrc/preferencesadminfrontend/controllers/CsvUploadControllerSpec.scala
@@ -17,11 +17,8 @@
 package uk.gov.hmrc.preferencesadminfrontend.controllers
 
 import org.apache.pekko.stream.Materializer
-import org.apache.pekko.stream.scaladsl.Framing.FramingException
-import org.mockito.ArgumentMatchers
 import org.mockito.ArgumentMatchers.*
 import org.mockito.Mockito.*
-import org.scalactic.source.Position
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatestplus.mockito.MockitoSugar
 import org.scalatestplus.play.guice.GuiceOneAppPerSuite
@@ -36,12 +33,9 @@ import play.api.test.FakeRequest
 import play.api.test.Helpers.*
 import play.api.{ Application, inject }
 import uk.gov.hmrc.http.HeaderCarrier
-import uk.gov.hmrc.preferencesadminfrontend.connectors.{ AlreadyOptedOut, OptedOut, PreferenceNotFound }
 import uk.gov.hmrc.preferencesadminfrontend.controllers.model.User
 import uk.gov.hmrc.preferencesadminfrontend.services.*
-import uk.gov.hmrc.preferencesadminfrontend.services.model.csv.UploadedBulKOptOutNinos
 import uk.gov.hmrc.preferencesadminfrontend.utils.SpecBase
-import uk.gov.hmrc.preferencesadminfrontend.views.messages.BulkUploadErrorMessages
 
 import java.nio.file.Path
 import scala.concurrent.{ ExecutionContext, Future }
@@ -85,8 +79,8 @@ class CsvUploadControllerSpec
     }
 
     "return 200 when service returns nil" in new TestCase {
-      val mockTemporaryFile = mock[TemporaryFile]
-      val mockPath = mock[Path]
+      val mockTemporaryFile: TemporaryFile = mock[TemporaryFile]
+      val mockPath: Path = mock[Path]
 
       when(mockTemporaryFile.path).thenReturn(mockPath)
 
@@ -118,391 +112,11 @@ class CsvUploadControllerSpec
     }
   }
 
-  "showBulkOptOutsUploadPage (GET /csv-upload-bulk-opt-outs)" should {
-    "return HTML" in new TestCase {
-      val result: Future[Result] =
-        controller.showBulkOptOutsUploadPage()(FakeRequest("", "").withSession(User.sessionKey -> "admin"))
-
-      contentType(result) mustBe Some("text/html")
-      charset(result) mustBe Some("utf-8")
-      status(result) mustBe OK
-
-      extractBulkOptOutErrors(contentAsString(result)) mustBe List.empty
-    }
-
-    "redirect to login page for non-admin user" in new TestCase {
-      val result: Future[Result] =
-        controller.showBulkOptOutsUploadPage()(FakeRequest("GET", "/decode"))
-      status(result) mustBe Status.SEE_OTHER
-
-    }
-  }
-
-  def extractBulkOptOutErrors(body: String): List[String] = {
-    val possibleErrorMessages = List(
-      BulkUploadErrorMessages.invalidFormatNinos,
-      BulkUploadErrorMessages.notFullyOptedIn,
-      BulkUploadErrorMessages.notFoundNinos,
-      BulkUploadErrorMessages.unexpectedFailureNinos,
-      BulkUploadErrorMessages.noUploadedNinos,
-      "Too many Ninos were uploaded",
-      BulkUploadErrorMessages.invalidFileFormat,
-      BulkUploadErrorMessages.duplicateNinos
-    )
-
-    possibleErrorMessages.collect {
-      case possibleError if body.contains(possibleError) =>
-        possibleError
-    }
-  }
-
-  "uploadBulkOptOuts (POST /csv-upload-bulk-opt-outs/confirmation)" should {
-    def filePart: FilePart[TemporaryFile] = FilePart(
-      key = "csvFile",
-      filename = "test.csv",
-      contentType = Some("text/csv"),
-      ref = mock[TemporaryFile]
-    )
-
-    "return 400 when form binding fails" in new TestCase {
-      val request = FakeRequest("", "")
-        .withSession(User.sessionKey -> "admin")
-
-      val result = controller.uploadBulkOptOuts()(request)
-
-      status(result) mustBe BAD_REQUEST
-    }
-
-    "return 200 but display a message saying the file could not be processed when it fails processing" in new TestCase {
-      val mockTemporaryFile = mock[TemporaryFile]
-      val mockPath = mock[Path]
-
-      when(mockTemporaryFile.path).thenReturn(mockPath)
-
-      when(mockBulkOptOutsService.readNinoBulkOptOutsFromFile(any())(any()))
-        .thenReturn(
-          Future.successful(
-            Left(
-              new FramingException("failed reading file")
-            )
-          )
-        )
-
-      val formData = MultipartFormData[TemporaryFile](
-        dataParts = Map.empty,
-        files = Seq(filePart),
-        badParts = Seq.empty
-      )
-
-      val request = FakeRequest("", "")
-        .withSession(User.sessionKey -> "admin")
-        .withBody(formData)
-
-      val result = controller.uploadBulkOptOuts()(request)
-
-      status(result) mustBe OK
-      val body: String = contentAsString(result)
-
-      extractBulkOptOutErrors(body) mustBe List(
-        "Invalid File Format. The uploaded file format could not be processed. Please enter a valid .csv file format."
-      )
-    }
-
-    "return 200 but display no entries were found if file was empty" in new TestCase {
-      val mockTemporaryFile = mock[TemporaryFile]
-      val mockPath = mock[Path]
-
-      when(mockTemporaryFile.path).thenReturn(mockPath)
-
-      when(mockBulkOptOutsService.readNinoBulkOptOutsFromFile(any())(any()))
-        .thenReturn(
-          Future.successful(Right(UploadedBulKOptOutNinos.empty))
-        )
-
-      val formData = MultipartFormData[TemporaryFile](
-        dataParts = Map.empty,
-        files = Seq(filePart),
-        badParts = Seq.empty
-      )
-
-      val request = FakeRequest("", "")
-        .withSession(User.sessionKey -> "admin")
-        .withBody(formData)
-
-      val result = controller.uploadBulkOptOuts()(request)
-
-      status(result) mustBe OK
-      val body: String = contentAsString(result)
-
-      extractBulkOptOutErrors(body) mustBe List(
-        BulkUploadErrorMessages.noUploadedNinos
-      )
-    }
-
-    "return 200 but display too many entries if the limit is exceeded" in new TestCase {
-      val mockTemporaryFile = mock[TemporaryFile]
-      val mockPath = mock[Path]
-
-      val ninos = (1 to 101).map { index =>
-        s"nino-$index"
-      }.toList
-
-      when(mockTemporaryFile.path).thenReturn(mockPath)
-
-      when(mockBulkOptOutsService.readNinoBulkOptOutsFromFile(any())(any()))
-        .thenReturn(
-          Future.successful(Right(UploadedBulKOptOutNinos.createUniqueFiltered(ninos)))
-        )
-
-      val formData = MultipartFormData[TemporaryFile](
-        dataParts = Map.empty,
-        files = Seq(filePart),
-        badParts = Seq.empty
-      )
-
-      val request = FakeRequest("", "")
-        .withSession(User.sessionKey -> "admin")
-        .withBody(formData)
-
-      val result = controller.uploadBulkOptOuts()(request)
-
-      status(result) mustBe OK
-      val body: String = contentAsString(result)
-
-      extractBulkOptOutErrors(body) mustBe List(
-        "Too many Ninos were uploaded"
-      )
-    }
-
-    def testValidParsing(
-      controller: CsvUploadController,
-      mockBulkOptOutsService: BulkUploadOptOutsService,
-      nonDistinctNinosFromCsvReading: List[String],
-      ninosThatWillbeProcessed: List[String],
-      serverResponse: List[BulkOptOutResult],
-      expectedErrors: List[String],
-      successCount: Int
-    )(implicit position: Position): Unit = {
-      val mockTemporaryFile = mock[TemporaryFile]
-      val mockPath = mock[Path]
-
-      when(mockTemporaryFile.path).thenReturn(mockPath)
-
-      val formData = MultipartFormData[TemporaryFile](
-        dataParts = Map.empty,
-        files = Seq(filePart),
-        badParts = Seq.empty
-      )
-
-      when(mockBulkOptOutsService.readNinoBulkOptOutsFromFile(any())(any()))
-        .thenReturn(
-          Future.successful(Right(UploadedBulKOptOutNinos.createUniqueFiltered(nonDistinctNinosFromCsvReading)))
-        )
-
-      when(
-        mockBulkOptOutsService
-          .processBulkOptOuts(ArgumentMatchers.eq(ninosThatWillbeProcessed))(any(), any(), any())
-      )
-        .thenReturn(
-          Future.successful(
-            serverResponse
-          )
-        )
-
-      val request = FakeRequest("", "")
-        .withSession(User.sessionKey -> "admin")
-        .withBody(formData)
-
-      val result = controller.uploadBulkOptOuts()(request)
-
-      status(result) mustBe OK
-      val body: String = contentAsString(result)
-
-      extractBulkOptOutErrors(body) mustBe expectedErrors
-      body must include(s"Successfully opted out $successCount records")
-    }
-
-    "return 200 when only valid entries that were at maximum upload count were parsed and processed successfully as opted out" in new TestCase {
-      val ninos = (1 to 100).map { index =>
-        s"nino-$index"
-      }.toList
-
-      testValidParsing(
-        controller = controller,
-        mockBulkOptOutsService = mockBulkOptOutsService,
-        nonDistinctNinosFromCsvReading = ninos,
-        ninosThatWillbeProcessed = ninos,
-        serverResponse = ninos.map(nino => ProcessedBulkOptOutResult(nino, OptedOut)),
-        expectedErrors = List.empty,
-        successCount = ninos.size
-      )
-    }
-
-    "return 200 when only valid entries were parsed but there were duplicates so these were separated out and used once" in new TestCase {
-      val nonDistinctNinos = List(
-        "nino1",
-        "nino1"
-      )
-
-      val distinctNinos = List(
-        "nino1"
-      )
-
-      testValidParsing(
-        controller = controller,
-        mockBulkOptOutsService = mockBulkOptOutsService,
-        nonDistinctNinosFromCsvReading = nonDistinctNinos,
-        ninosThatWillbeProcessed = distinctNinos,
-        serverResponse = distinctNinos.map(nino => ProcessedBulkOptOutResult(nino, OptedOut)),
-        expectedErrors = List("The following uploaded Ninos were duplicates, so were attempted once"),
-        successCount = 1
-      )
-    }
-
-    "return 200 when only valid entries were parsed but were already opted out" in new TestCase {
-      val ninos = List(
-        "nino1",
-        "nino2"
-      )
-
-      testValidParsing(
-        controller = controller,
-        mockBulkOptOutsService = mockBulkOptOutsService,
-        nonDistinctNinosFromCsvReading = ninos,
-        ninosThatWillbeProcessed = ninos,
-        serverResponse = List(
-          ProcessedBulkOptOutResult("nino1", AlreadyOptedOut),
-          ProcessedBulkOptOutResult("nino2", AlreadyOptedOut)
-        ),
-        expectedErrors = List(BulkUploadErrorMessages.notFullyOptedIn),
-        successCount = 0
-      )
-    }
-
-    "return 200 when only valid entries were parsed but were not found" in new TestCase {
-      val ninos = List(
-        "nino1",
-        "nino2"
-      )
-
-      testValidParsing(
-        controller = controller,
-        mockBulkOptOutsService = mockBulkOptOutsService,
-        nonDistinctNinosFromCsvReading = ninos,
-        ninosThatWillbeProcessed = ninos,
-        serverResponse = List(
-          ProcessedBulkOptOutResult("nino1", PreferenceNotFound),
-          ProcessedBulkOptOutResult("nino2", PreferenceNotFound)
-        ),
-        expectedErrors = List(BulkUploadErrorMessages.notFoundNinos),
-        successCount = 0
-      )
-    }
-
-    "return 200 when only valid entries were parsed but failed for unknown reasons" in new TestCase {
-      val ninos = List(
-        "nino1",
-        "nino2"
-      )
-
-      testValidParsing(
-        controller = controller,
-        mockBulkOptOutsService = mockBulkOptOutsService,
-        nonDistinctNinosFromCsvReading = ninos,
-        ninosThatWillbeProcessed = ninos,
-        serverResponse = List(
-          FailedCallBulkOptOutResult("nino1"),
-          FailedCallBulkOptOutResult("nino2")
-        ),
-        expectedErrors = List(BulkUploadErrorMessages.unexpectedFailureNinos),
-        successCount = 0
-      )
-    }
-
-    "return 200 when only valid entries were parsed but failed due to an invalid formatted Nino value" in new TestCase {
-      val ninos = List(
-        "nino1",
-        "nino2"
-      )
-
-      testValidParsing(
-        controller = controller,
-        mockBulkOptOutsService = mockBulkOptOutsService,
-        nonDistinctNinosFromCsvReading = ninos,
-        ninosThatWillbeProcessed = ninos,
-        serverResponse = List(
-          InvalidNinoBulkOptOutResult("nino1"),
-          InvalidNinoBulkOptOutResult("nino2")
-        ),
-        expectedErrors = List(BulkUploadErrorMessages.invalidFormatNinos),
-        successCount = 0
-      )
-    }
-
-    "return 200 and display all combinations of reponses" in new TestCase {
-      val ninos = List(
-        "nino1",
-        "nino2",
-        "nino3",
-        "nino4",
-        "nino4",
-        "nino5",
-        "nino6",
-        "nino7",
-        "nino8",
-        "nino8",
-        "nino9"
-      )
-
-      val disinctNinos = List(
-        "nino1",
-        "nino2",
-        "nino3",
-        "nino4",
-        "nino5",
-        "nino6",
-        "nino7",
-        "nino8",
-        "nino9"
-      )
-
-      testValidParsing(
-        controller = controller,
-        mockBulkOptOutsService = mockBulkOptOutsService,
-        nonDistinctNinosFromCsvReading = ninos,
-        ninosThatWillbeProcessed = disinctNinos,
-        serverResponse = List(
-          ProcessedBulkOptOutResult("nino1", PreferenceNotFound),
-          ProcessedBulkOptOutResult("nino2", OptedOut),
-          FailedCallBulkOptOutResult("nino3"),
-          FailedCallBulkOptOutResult("nino3"),
-          FailedCallBulkOptOutResult("nino4"),
-          ProcessedBulkOptOutResult("nino5", AlreadyOptedOut),
-          ProcessedBulkOptOutResult("nino6", AlreadyOptedOut),
-          ProcessedBulkOptOutResult("nino7", OptedOut),
-          ProcessedBulkOptOutResult("nino8", OptedOut),
-          InvalidNinoBulkOptOutResult("nino9")
-        ),
-        expectedErrors = List(
-          BulkUploadErrorMessages.invalidFormatNinos,
-          BulkUploadErrorMessages.notFullyOptedIn,
-          BulkUploadErrorMessages.notFoundNinos,
-          BulkUploadErrorMessages.unexpectedFailureNinos,
-          BulkUploadErrorMessages.duplicateNinos
-        ),
-        successCount = 3
-      )
-    }
-
-  }
-
   class TestCase {
     lazy val mockUploadService: UploadService = mock[UploadService]
-    lazy val mockBulkOptOutsService: BulkUploadOptOutsService = mock[BulkUploadOptOutsService]
 
     implicit lazy val app: Application = GuiceApplicationBuilder()
       .overrides(inject.bind[UploadService].toInstance(mockUploadService))
-      .overrides(inject.bind[BulkUploadOptOutsService].toInstance(mockBulkOptOutsService))
       .build()
 
     implicit val hc: HeaderCarrier = HeaderCarrier()
